requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
click==8.3.1
fastapi==0.129.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
playwright==1.58.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dotenv==1.2.1
PyYAML==6.0.3
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
